+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/response/JsonRpcErrorResponse.java
@@ -89,14 +89,19 @@ public class JsonRpcErrorResponse implements JsonRpcResponse {
       final String encodedReasonText =
           revertReason.toHexString().substring(errorMethodABI.length());
 
-      List<TypeReference<Type>> revertReasonTypes =
-          Collections.singletonList(TypeReference.create((Class<Type>) AbiTypes.getType("string")));
-      List<Type> decoded = FunctionReturnDecoder.decode(encodedReasonText, revertReasonTypes);
+      try {
+        List<TypeReference<Type>> revertReasonTypes =
+            Collections.singletonList(
+                TypeReference.create((Class<Type>) AbiTypes.getType("string")));
+        List<Type> decoded = FunctionReturnDecoder.decode(encodedReasonText, revertReasonTypes);
 
-      // Expect a single decoded string
-      if (decoded.size() == 1 && (decoded.get(0) instanceof Utf8String)) {
-        Utf8String decodedRevertReason = (Utf8String) decoded.get(0);
-        return Optional.of(decodedRevertReason.getValue());
+        // Expect a single decoded string
+        if (decoded.size() == 1 && (decoded.get(0) instanceof Utf8String)) {
+          Utf8String decodedRevertReason = (Utf8String) decoded.get(0);
+          return Optional.of(decodedRevertReason.getValue());
+        }
+      } catch (StringIndexOutOfBoundsException exception) {
+        return Optional.of("ABI decode error");
       }
     }
     return Optional.empty();