+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Breaking Changes
 
 - Removed support for version 0 of the database as it is no longer used by any active node.
+- Add ABI-decoded revert reason to `eth_call` and `eth_estimateGas` responses [#5705](https://github.com/hyperledger/besu/issues/5705)
 
 ### Additions and Improvements
 - EvmTool now executes the `execution-spec-tests` via the `t8n` and `b11r`. See the [README](ethereum/evmtool/README.md) in EvmTool for more instructions.